+++ src/app/features/groups/organisms/types/GroupTableTypes.ts
@@ -1,0 +1,11 @@
+export type Group = {
+  id: string;
+  name: string;
+  user_count: number;
+  drone_count: number;
+};
+
+export type GroupTableTypes = {
+  group: Group[];
+  handleRedirectDetailGroup: (id: string) => void;
+};
